requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pillow==10.4.0
onnxruntime==1.19.2
numpy==1.26.4
